# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Brew Haven ERP API
  version: 1.0.0
servers:
  - url: http://localhost:4000
paths:
  /api/v1/auth/login:
    post:
      summary: Authenticate a user and issue a JWT
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        "200":
          description: Authenticated
  /api/v1/inventory/items:
    get:
      summary: List inventory items
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Inventory returned
    post:
      summary: Create an inventory item
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Inventory item created
  /api/v1/sales/orders:
    get:
      summary: List sales orders
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Orders returned
    post:
      summary: Create a sales order
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Sales order created
  /api/v1/purchasing/suppliers:
    get:
      summary: List suppliers
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Suppliers returned
  /api/v1/purchasing/purchase-orders:
    get:
      summary: List purchase orders
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Purchase orders returned
    post:
      summary: Create a purchase order
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Purchase order created
  /api/v1/crm/customers:
    get:
      summary: List customers
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Customers returned
    post:
      summary: Create a customer
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Customer created
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT